+++ src/hooks.ts
@@ -32,7 +32,6 @@ Before(async function () {
         ...driverConfig.timeout
     }
     config.driverConfig = driverConfig;
-    console.log(driverConfig)
     global.browser = global.browser ? global.browser : await driverProvider(config.driverConfig);
     global.context = await browser.newContext(config?.driverConfig?.capabilities);
     if (config.driverConfig.trace) {