+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.25.0]
 - :rocket: added enabled/disabled validation
 - :beetle: fixed issue with default timeouts
-
+- :beetle: fixed regexp in dynamic po steps
+  
 ## [0.0.24]
 - :rocket: added _I open new tab_
 - :rocket: added steps to work with multiple browser contexts